blocks: use a hex-string instead of a placeholder for binary data
This improves throughput by 12% by reducing the round-trips to execute a
request. When using placeholders we have to prepare, then
bind+execute. Found when trying to figure out how cassandra and mongo
were executing requests faster than we could parse them (without
executing). This is an effort to make the comparison with
cassandra/mongo as apples-to-apples as possible.

Requires https://github.com/cockroachdb/cockroach/pull/13287