YCSB: Increment write count only after insertion
This was not showing up in single node testing, but in a distributed
cockroach cluster, there is a small probability that write() writes
key=N, and the next read call draws N, and dispatches it to a
different cockroach cluster before the write() is committed, resulting
in a read error. Therefore maintain two counters, iMax and
iMaxHead. iMaxHead is used for handing out write keys, while iMax is
the upper bound for read keys.

increment the global write counter only
after write() has committed.