kv: clarify generator semantics
The previous statement that read keys are guaranteed to
exist seems wrong. Since a common sequence is shared,
all that is being guaranteed is that if a read is issued
on one generator, there must have been a write issued on
any generator (not necessarily this one) that wrote a
key for the sequence number from which the hash was
created. Not exactly useful. Similarly, write keys may
already exist, either due to hash collisions, or due to
use of the `--cycle-length` parameter.