tpcc: index hints and direction improvements
We should use a descending index on order.o_id, since we always use that
key component in a reverse scan. This improves performance by a lot.

Also, hint that the customer_idx index should be used when looking up by
last name. It's not clear why the optimizer didn't notice this by
itself. Presumably it thought the cost of the index join was too high.