loadgen: add load generator that runs and profiles queries given
Queries are provided in a file that can be passed to the load generator.
The load generator then runs these queries and reports the latency of
each at the end of execution.

This load generator is useful to be able to answer questions like "what
is the latency of this set of queries using distsql? How about without
distsql?" easily.