tpcc: add thinking/keying time and cap workers
This commit corrects the tpcc implementation to be to spec with regards
to concurrency and tpmC.

Specifically, all workers now have a keying/thinking time sleep
before/after the actual operation. These sleeps depend on the operation
being performed and are coded to the tpcc spec.

Also, each warehouse now has a maximum of 10 workers as dictated by the
spec. This means that the tpmC of a tpcc load test should be between 8
and 12 tpmC per warehouse, which requires any interesting test with tpcc
to have a lot more data than we've been testing with.

You can disable this new functionality with the -no-wait flag, which
removes the thinking/keying time and cap on concurrency.