Always bind workers to warehouses
Change the behavior when `-no-wait` is specified to assign each worker
to a warehouse. Previously each worker was selecting a random warehouse
to operate on. This allows experimentation with contention. For example,
`-warehouses=1 -concurrency=1 -no-wait` shows the throughput of a single
worker without contention. Specify higher values of concurrency to see
more contention. And we can also specify `-warehouses=10
-concurrency=10` to see scalability in the absence of contention.

Change the ramp-up behavior to factor in the think time as well. This is
an attempt to reduce the thundering herd effect we're seeing when the
load generator starts.