tpcc: use deck method of regulating txn mix
The naive randomness used originally (picking randomly from the
transaction mix distribution without hysteris) leads to a potentially
invalid transaction mix.

The spec suggests using a deck of transactions to regulate the mix,
where the deck is iterated through in random order and shuffled when
exhausted. Now, we use that method.

Note that the default transaction mix changed slightly. The spec isn't
precise on what the transaction mix is, but it clearly says that the
deck method can be implemented with 10 cards for each of payment and new
order, and 1 each of the rest, so that's what the mix is now.