tpcc: add -partition flag
Partition all of the tables and indexes. Most of the tables are
partitioned by warehouse ID. The item table is partitioned by item
ID. The primary index of the history table is partitioned by
rowid (random 63-bit positive integer) while the other history table
indexes are partitioned by warehouse ID.

The motivation here is to be able to spread the partitions evenly in a
cluster in order to achieve better range distribution. This will require
setting zone configs for all of these partitions and starting up the
cluster with attributes or localities on each node.