tpcc: add auditor and initial check
I added some basic auditing functionality to the TPC-C workload. This is
similar to the existing checks in checks.go, but rather than querying
the db it relies on statistics which are gathered as the workload runs.
Right now the audit checks are performed only when the workload exits,
but in the future we could also consider running them periodically as it
runs.

Refers #151