tpcc: fix random range of several events
rand.Intn is exclusive on the right, not inclusive. This caused several
mistakes in implementation.

1. `11.11...%` of customers in the initial dataset have bad credit
instead of 10%. this causes the payment transaction to do slightly more
work - when customers have bad credit, some strings must be appended and
returned to the user.

2. the read-only orderStatus transaction was only selecting over 9
districts, instead of 10. this will slightly reduce contention when
fixed.

3. `11.11...%` of items have the string ORIGINAL in them instead of 10%.
this has no appreciable effect.

4. the read-only stockLevel transaction only selected over 9 districts,
instead of 10. this will slightly reduce contention when fixed.
```